requirements.txt
# Requirements
# Install with: pip install -r requirements.txt

# Core
ollama>=0.1.0
pydantic>=2.0.0
rich>=13.0.0
typer>=0.9.0
httpx>=0.25.0
pyyaml>=6.0
python-dotenv>=1.0.0

# Optional - for development
pytest>=7.0.0
pytest-asyncio>=0.21.0
black>=23.0.0
ruff>=0.1.0